Refactor useChart.ts: Split hooks into separate files
- Extracted `useChartProcessing`, `useChartOptimization`, `useChartA11y` from `useChart.ts`
- Renamed conflicting `useChartData.ts` to `useSimpleChartData.ts`
- Restored original `useChartPerformance.ts` logic
- Updated `ChartRenderer` and exports to use new hook names
- Updated `src/lib/composables/index.ts` to export `useSimpleChartData`
